+++ scripts/run_vsim.sh
@@ -29,7 +29,7 @@ SEEDS=(0)
 
 call_vsim() {
     for seed in ${SEEDS[@]}; do
-        echo "run -all" | $VSIM -sv_seed $seed "$@" | tee vsim.log 2>&1
+        echo "run -all" | $VSIM -sv_seed $seed "$@" | tee vsim.log 2>&1 #-gui
         grep "Errors: 0," vsim.log
     done
 }
@@ -174,6 +174,24 @@ exec_test() {
                 done
             done
             ;;
+        axi_xp)
+            for NumMst in 1 4; do
+                for NumSlv in 1 4; do
+                    for Atop in 0 1; do
+                        for Exclusive in 0 1; do
+                            for UniqueIds in 0 1; do
+                                call_vsim tb_axi_xp -gTbNumMst=$NumMst -gTbNumSlv=$NumSlv \
+                                        -gTbEnAtop=$Atop -gTbEnExcl=$Exclusive \
+                                        -gTbUniqueIds=$UniqueIds
+                            done
+                        done
+                    done
+                done
+            done
+            ;;
+        axi_dma_backend)
+            call_vsim tb_$1 -voptargs="+acc +cover=bcesfx"
+            ;; 
         *)
             call_vsim tb_$1 -t 1ns -coverage -voptargs="+acc +cover=bcesfx"
             ;;